rename useFeatureFlag=>useDynamicConfig & fix exports
